+++ extensions/Worker.Extensions.CosmosDB/src/Config/CosmosDBBindingOptionsSetup.cs
@@ -16,14 +16,14 @@ namespace Microsoft.Azure.Functions.Worker
         private readonly IConfiguration _configuration;
         private readonly AzureComponentFactory _componentFactory;
         private readonly IOptionsMonitor<WorkerOptions> _workerOptions;
-        private readonly IOptionsMonitor<CosmosDBOptions> _cosmosDBOptions;
+        private readonly IOptionsMonitor<CosmosDBExtensionOptions> _cosmosExtensionOptions;
 
-        public CosmosDBBindingOptionsSetup(IConfiguration configuration, AzureComponentFactory componentFactory, IOptionsMonitor<WorkerOptions> workerOptions, IOptionsMonitor<CosmosDBOptions> cosmosOptions)
+        public CosmosDBBindingOptionsSetup(IConfiguration configuration, AzureComponentFactory componentFactory, IOptionsMonitor<WorkerOptions> workerOptions, IOptionsMonitor<CosmosDBExtensionOptions> cosmosExtensionOptions)
         {
             _configuration = configuration ?? throw new ArgumentNullException(nameof(configuration));
             _componentFactory = componentFactory ?? throw new ArgumentNullException(nameof(componentFactory));
             _workerOptions = workerOptions ?? throw new ArgumentNullException(nameof(workerOptions));
-            _cosmosDBOptions = cosmosOptions ?? throw new ArgumentNullException(nameof(cosmosOptions));
+            _cosmosExtensionOptions = cosmosExtensionOptions ?? throw new ArgumentNullException(nameof(cosmosExtensionOptions));
         }
 
         public void Configure(CosmosDBBindingOptions options)
@@ -60,7 +60,7 @@ namespace Microsoft.Azure.Functions.Worker
             }
 
             options.Serializer = new WorkerCosmosSerializer(_workerOptions.CurrentValue.Serializer);
-            options.CosmosDBOptions = _cosmosDBOptions.CurrentValue;
+            options.CosmosExtensionOptions = _cosmosExtensionOptions.CurrentValue;
         }
     }
 }